requirements/main.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/main.in
#
aiodns==2.0.0             # via -r requirements/main.in
aiohttp==3.6.2            # via -r requirements/main.in, aiojobs, pytest-aiohttp
aiojobs==0.2.2            # via -r requirements/main.in
aiopg[sa]==1.0.0          # via -r requirements/main.in
alembic==1.4.2            # via -r requirements/main.in
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp, pytest
cchardet==2.1.6           # via -r requirements/main.in
cffi==1.14.0              # via pycares
chardet==3.0.4            # via aiohttp
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
idna==2.9                 # via yarl
imageio==2.8.0            # via scikit-image
joblib==0.15.1            # via scikit-learn
kiwisolver==1.2.0         # via matplotlib
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via mako
marshmallow==3.6.1        # via -r requirements/main.in
matplotlib==3.2.2         # via scikit-image
more-itertools==8.4.0     # via pytest
multidict==4.7.6          # via aiohttp, yarl
networkx==2.4             # via scikit-image
numpy==1.18.5             # via imageio, matplotlib, pywavelets, scikit-image, scikit-learn, scipy, tifffile
packaging==20.4           # via pytest
pillow==7.1.2             # via -r requirements/main.in, imageio, scikit-image
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via aiopg, sqlalchemy
py==1.8.2                 # via pytest
pycares==3.1.1            # via aiodns
pycparser==2.20           # via cffi
pyparsing==2.4.7          # via matplotlib, packaging
pytest-aiohttp==0.3.0     # via -r requirements/main.in
pytest==5.4.3             # via -r requirements/main.in, pytest-aiohttp
python-dateutil==2.8.1    # via alembic, matplotlib
python-dotenv==0.13.0     # via -r requirements/main.in
python-editor==1.0.4      # via alembic
pywavelets==1.1.1         # via scikit-image
pyyaml==5.3.1             # via -r requirements/main.in
scikit-image==0.17.2      # via -r requirements/main.in
scikit-learn==0.23.1      # via -r requirements/main.in
scipy==1.4.1              # via scikit-image, scikit-learn
six==1.15.0               # via cycler, packaging, python-dateutil, tenacity
sqlalchemy[postgresql_psycopg2binary]==1.3.17  # via aiopg, alembic
tenacity==6.2.0           # via -r requirements/main.in
threadpoolctl==2.1.0      # via scikit-learn
tifffile==2020.6.3        # via scikit-image
wcwidth==0.2.4            # via pytest
yarl==1.4.2               # via aiohttp
